implemented the auth

+++ Onboarding/ecommerce-clean-architecture/lib/features/product/domain/entities/product.dart
@@ -1,3 +1,4 @@
+import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
 
 class Productentity extends Equatable{
@@ -5,7 +6,6 @@ const Productentity({
   required this.id,
   required this.image,
   required this.name,
-  // required this.type,
   required this.description,
   required this.price,
 }
@@ -13,16 +13,15 @@ const Productentity({
   final String id;
   final String image;
   final String name;
-  // final String type;
   final String description;
-  final int price;
+  final double price;
+
 
   @override
   List<Object ? > get props=>[
     id,
     image,
     name,
-    // type,
     description,
     price,
   ];
